Add REST API to js-sdk
